Separated hooks into separate files and fixed some types
