+++ src/ui/context/hooks/useDeskproAppClient.ts
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+import { DeskproAppContext } from "../context";
+import { DeskproAppClient, DeskproAppContextValue } from "../types";
+
+export const useDeskproAppClient = (): DeskproAppClient => {
+  const value = useContext<DeskproAppContextValue>(DeskproAppContext);
+
+  return {
+    client: value?.client ?? null,
+  };
+};
